pyproject.toml
[project]
name = "agent-orchestrator"
version = "0.1.0"
description = "Multi-agent orchestration system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ollama>=0.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "ddgs>=1.0",
]

[project.optional-dependencies]
claude = [
    "anthropic>=0.18.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents"]